Refactor reload event into plugin package
