+++ core/src/main/java/com/clubobsidian/dynamicgui/core/command/DynamicGuiCommand.java
@@ -22,7 +22,7 @@ import cloud.commandframework.annotations.CommandPermission;
 import com.clubobsidian.dynamicgui.core.DynamicGui;
 import com.clubobsidian.dynamicgui.core.Key;
 import com.clubobsidian.dynamicgui.core.entity.PlayerWrapper;
-import com.clubobsidian.dynamicgui.core.event.DynamicGuiReloadEvent;
+import com.clubobsidian.dynamicgui.core.event.plugin.DynamicGuiReloadEvent;
 import com.clubobsidian.dynamicgui.core.gui.Gui;
 import com.clubobsidian.dynamicgui.core.manager.dynamicgui.GuiManager;
 import com.clubobsidian.dynamicgui.core.platform.Platform;
